Refactor and use cipherStorage Base class
